+++ src/main/get-cursor-node.js
@@ -1,5 +1,5 @@
+import { getDescendants } from "../utils/ast-utils.js";
 import createGetVisitorKeysFunction from "./create-get-visitor-keys-function.js";
-import { getDescendants } from "./ast-utils.js";
 
 function getCursorNode(ast, options) {
   const { cursorOffset, locStart, locEnd } = options;
